modules/kafka-producer/requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.1
kafka-python==2.0.2
protobuf==4.22.0
